[Day 1 :: Drum Kit] Added index.html
The index.html file contains code to import the CSS stylesheets and the Javascript source codes. It also has a set of DIV and nested SPAN tags that represent various keys displayed on the page. Use has been made of Bootstrap to make the page eresponsive. Further styles for elements can be found in the stylesheet