[Day 1 :: Drum Kit] Added stylesheet
The styles included in the stylesheet, style.css, describe the layout of the web page. It adds the background image, styles for the keys, positioning of various element, etc. Also, the class key-pressed, which is used to animate a key element when the corressponding element is pressed is defined here.