[Day 1: Drum Kit] Post Event Handling
After the event has been triggered and handled, the key element that was affected needs to be restored to its original value. The patch provided here is responsible for detecting whether the animation is over, and if over, restore the styles of the element. This is done by removing the key-pressed class from the element.